+++ CHANGELOG.rst
@@ -15,6 +15,12 @@ recent versions
     with ``get_source_field`` providing a ``src`` that consists of x-, y-,
     and z-coordinates of all endpoints of the individual segments.
 
+  - Simple magnetic "dipole" sources can now be created with
+    ``get_source_field`` by providing a point dipole (``[x, y, z, azm, dip]``)
+    and set ``msrc=True`` (new fct-keyword). This will create a square loop of
+    1x1 m perpendicular to the defined point dipole, hence simulating a
+    magnetic source.
+
 
 v0.15.3: Move to EMSiG
 ----------------------